Use decimal to match other meal endpoints and entity. (#8)

+++ src/CgmLink.Api/Endpoints/Meals/UpdateMeal/UpdateMealRequest.cs
@@ -24,7 +24,7 @@ public sealed record NewMealIngredientRequest
 {
     public Guid? Id { get; set; }
     public required Guid IngredientId { get; set; }
-    public required int Quantity { get; set; }
+    public required decimal Quantity { get; set; }
 
     public sealed class NewMealIngredientValidator : AbstractValidator<NewMealIngredientRequest>
     {
